// frontend/src/types/generated/mobile/conectaPROMobileAPI.schemas.ts
export type DeviceTokenCreateDeviceId = string | null;

export type DeviceTokenCreateDeviceName = string | null;

export type DeviceTokenCreateDeviceModel = string | null;

export type DeviceTokenCreateOsVersion = string | null;

export type DeviceTokenCreateAppVersion = string | null;

export type DeviceTokenCreateAppBuild = string | null;

export type DeviceTokenCreateTimezone = string | null;

/**
 * Schema para criar/registrar token de dispositivo.
 */
export interface DeviceTokenCreate {
  /**
   * Token FCM/APNs
   * @minLength 10
   * @maxLength 500
   */
  token: string;
  /** android, ios, web */
  platform: string;
  device_id?: DeviceTokenCreateDeviceId;
  device_name?: DeviceTokenCreateDeviceName;
  device_model?: DeviceTokenCreateDeviceModel;
  os_version?: DeviceTokenCreateOsVersion;
  app_version?: DeviceTokenCreateAppVersion;
  app_build?: DeviceTokenCreateAppBuild;
  /** @maxLength 10 */
  locale?: string;
  timezone?: DeviceTokenCreateTimezone;
}